requirement.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
flask-blueprint==1.3.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.15.0
MarkupSafe==1.1.1
pycparser==2.19
python-dotenv==0.10.3
requests==2.22.0
SQLAlchemy==1.3.12
urllib3==1.25.7
Werkzeug==0.16.0
